i kill the leaks

+++ Violet/Maps/Map.cs
@@ -36,11 +36,9 @@ namespace Violet.Maps
             {
                 arg = this.Head.Tilesets[0].Name;
             }
-
             string resource = string.Format("{0}{1}.dat", graphicDirectory, arg);
             IList<TileGroup> list = new List<TileGroup>(this.Groups.Count);
             long ticks = DateTime.Now.Ticks;
-            //List<Tile> globalTiles = new List<Tile>(); 
             for (int i = 0; i < this.Groups.Count; i++)
             {
                 Map.Group group = this.Groups[i];
@@ -49,7 +47,7 @@ namespace Violet.Maps
                 int num = 0;
                 while (j < group.Tiles.Length)
                 {
-                    int num2 = group.Tiles[j] - 1;
+                    int num2 = (int)(group.Tiles[j] - 1);
                     if (num2 >= 0)
                     {
                         ushort num3;
@@ -62,23 +60,20 @@ namespace Violet.Maps
                             num3 = 0;
                         }
                         int num4 = group.Width * 8;
-                        Vector2f position = new Vector2f(num * 8L % num4, num * 8L / num4 * 8L);
+                        Vector2f position = new Vector2f((float)((long)num * 8L % (long)num4), (float)((long)num * 8L / (long)num4 * 8L));
                         bool flipHoriz = (num3 & 1) > 0;
                         bool flipVert = (num3 & 2) > 0;
                         bool flipDiag = (num3 & 4) > 0;
                         ushort animId = (ushort)(num3 >> 3);
                         Tile item = new Tile((uint)num2, position, flipHoriz, flipVert, flipDiag, animId);
                         list2.Add(item);
-                       // globalTiles.Add(item);
                     }
                     j += 2;
                     num++;
                 }
-                TileGroup item2 = new TileGroup(list2, resource, group.Depth, new Vector2f(group.X, group.Y), palette);
+                TileGroup item2 = new TileGroup(list2, resource, group.Depth, new Vector2f((float)group.X, (float)group.Y), palette);
                 list.Add(item2);
             }
-            //globalTilelist = globalTiles;
-            //Debug.LDebug($"global tile list count: {globalTilelist.Count}");
             Debug.LInfo($"Created tile groups in {(DateTime.Now.Ticks - ticks) / 10000L}ms");
             return list;
         }
